# swagger-user.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ember API 서버",
...
    "version": "v2.2"
  },
  "servers": [
    {
      "url": "https://ember-app.duckdns.org",
      "description": "Ember 배포 서버"
    }
  ],
...
  "paths": {
...
    "/api/matching/{diaryId}/skip": {
      "post": {
        "tags": [
          "Matching"
        ],
        "summary": "추천 일기 넘기기 (7일 재추천 제외)",
        "description": "해당 일기를 넘깁니다.\n\n**동작:** matching_pass 테이블에 기록, 7일간 탐색에서 재노출 방지",
        "operationId": "skip",
        "parameters": [
          {
            "name": "diaryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": null
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/matching/{diaryId}/select": {
      "post": {
        "tags": [
          "Matching"
        ],
        "summary": "교환일기 신청 (상대 일기 선택)",
        "description": "해당 일기 작성자에게 교환일기를 신청합니다.\n\n**동작:**\n- PENDING 상태로 매칭 요청 생성\n- 상대방에게 MATCHING_REQUEST 알림 발송\n- **양방향 매칭 감지:** 상대도 나에게 신청했으면 자동으로 매칭 성사 → 교환일기 방 생성\n- 비관적 락(PESSIMISTIC_WRITE)으로 Race Condition 방지\n\n**에러:** M001(이미 신청), M003(차단), M004(일기 없음), M005(교환 3건 만석)",
        "operationId": "select",
        "parameters": [
          {
            "name": "diaryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "신청 성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": {
                    "matchingId": 1,
                    "isMatched": false,
                    "roomUuid": null
                  }
                }
              }
            }
          },
          "400": {
            "description": "자기 일기에 신청 (M005)",
            "content": {
              "application/json": {
                "example": {
                  "code": "M005",
                  "message": "본인의 일기에는 매칭 요청할 수 없습니다",
                  "data": null
                }
              }
            }
          },
          "409": {
            "description": "이미 신청한 일기 (M001)",
            "content": {
              "application/json": {
                "example": {
                  "code": "M001",
                  "message": "이미 매칭 요청을 보낸 상대입니다",
                  "data": null
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/matching/requests/{matchingId}/accept": {
      "post": {
        "tags": [
          "Matching"
        ],
        "summary": "매칭 요청 수락 (매칭 성사)",
        "description": "매칭 요청을 수락합니다.\n\n**동작:**\n- 매칭 상태 PENDING → MATCHED\n- 교환일기 방 자동 생성 (roomUuid 반환)\n- 양측에게 MATCHING_MATCHED 알림 발송\n\n**에러:** M006(존재하지 않는 매칭)",
        "operationId": "acceptRequest",
        "parameters": [
          {
            "name": "matchingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "수락 성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": {
                    "matchingId": 1,
                    "isMatched": true,
                    "roomUuid": "bf48bc80-4773-4d50-a293-3ba5f7cea823"
                  }
                }
              }
            }
          },
          "404": {
            "description": "매칭 없음 (M006)",
            "content": {
              "application/json": {
                "example": {
                  "code": "M006",
                  "message": "매칭 요청을 찾을 수 없습니다",
                  "data": null
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/matching/requests": {
      "get": {
        "tags": [
          "Matching"
        ],
        "summary": "받은 매칭 요청 목록 조회",
        "description": "받은 매칭 요청 목록을 조회합니다.\n\n**응답:** fromUserNickname, fromUserAgeGroup, 일기 미리보기 포함\n- matchingId를 POST /api/matching/requests/{matchingId}/accept에 전달",
        "operationId": "getReceivedRequests",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": [
                    {
                      "matchingId": 1,
                      "fromUserId": 5,
                      "fromUserNickname": "미소짓는풍선",
                      "fromUserAgeGroup": "20대",
                      "diaryId": 10,
                      "diaryPreview": "오늘은...",
                      "requestedAt": "2026-04-30T10:00:00"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/matching/recommendations": {
      "get": {
        "tags": [
          "Matching"
        ],
        "summary": "AI 기반 추천 일기 목록",
        "description": "AI 기반 매칭 추천 목록을 조회합니다.\n\n**동작:**\n- KoSimCSE 유사도 기반 상위 추천\n- Redis 캐시(24h) → 만료 시 stale 캐시 폴백 (빈 응답 방지)\n- AI FastAPI 미연동 시 AI003 에러\n\n**에러:** AI003(AI 매칭 서버 오류)",
        "operationId": "getRecommendations",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": {
                    "generatedAt": "2026-04-30T10:00:00",
                    "source": "FRESH",
                    "items": [
                      {
                        "userId": 5,
                        "matchingScore": 0.85,
                        "breakdown": {
                          "keywordOverlap": 1,
                          "cosineSimilarity": 0.85
                        },
                        "summary": null
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/matching/recommendations/{diaryId}/preview": {
      "get": {
        "tags": [
          "Matching"
        ],
        "summary": "추천 일기 블라인드 미리보기",
        "description": "추천 일기의 블라인드 미리보기를 조회합니다.\n\n**응답:**\n- 본문 앞 100자만 노출\n- 유사도 배지: 0.7이상 \"잘 맞을 것 같아요\", 0.5~0.7 \"공통점이 있어요\"",
        "operationId": "preview_1",
        "parameters": [
          {
            "name": "diaryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": {
                    "diaryId": 10,
                    "preview": "오늘은...",
                    "similarityBadge": "잘 맞을 것 같아요"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/matching/lifestyle-report": {
      "get": {
        "tags": [
          "Matching"
        ],
        "summary": "라이프스타일 분석 리포트",
        "description": "내 라이프스타일 리포트를 조회합니다.\n\n**활성화 조건:** 일기 5편 이상 작성\n- 5편 미만이면 analysisAvailable=false, requiredDiaryCount=5 반환\n\n**응답 (활성화 시):**\n- 활동 히트맵, 요일별 패턴, 감정 그래프, 평균 일기 길이 등",
        "operationId": "lifestyleReport",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": {
                    "analysisAvailable": true,
                    "requiredDiaryCount": 5,
                    "currentDiaryCount": 7
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/diaries/{diaryId}/detail": {
      "get": {
        "tags": [
          "Matching"
        ],
        "summary": "탐색 일기 상세 조회 (유사도 배지 + 작성자 다른 일기)",
        "description": "탐색에서 선택한 일기의 상세를 조회합니다.\n\n**응답 포함:**\n- 일기 전체 본문\n- 작성자 성격 키워드 (AI 분석 결과)\n- 작성자의 다른 공개 일기 미리보기 (최대 3건)\n- 유사도 점수/배지 (AI 연동 시)",
        "operationId": "diaryDetail",
        "parameters": [
          {
            "name": "diaryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": {
                    "diaryId": 10,
                    "authorId": 5,
                    "ageGroupLabel": "20대",
                    "content": "오늘은...",
                    "summary": "AI 요약",
                    "keywords": [
                      "안정 추구"
                    ],
                    "moodTags": [
                      "감성적인"
                    ],
                    "category": "DAILY",
                    "createdAt": "2026-04-30T10:00:00",
                    "similarityBadge": null,
                    "otherDiariesPreview": [
                      {
                        "diaryId": 11,
                        "summary": "요약",
                        "createdAt": "2026-04-29T10:00:00"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/diaries/explore": {
      "get": {
        "tags": [
          "Matching"
        ],
        "summary": "일기 탐색 (커서 기반 페이징, 정렬/필터 지원)",
        "description": "다른 사용자의 일기를 탐색합니다 (카드 스와이프 UI).\n\n**쿼리 파라미터:**\n- `cursor` (선택): 이전 응답의 nextCursor 값\n- `size` (기본 10): 한 번에 가져올 개수\n- `sort` (기본 latest): `latest`(최신순)\n- `sido` (선택): 시/도 필터 (예: \"서울특별시\")\n- `sigungu` (선택): 시/군/구 필터\n- `ageGroup` (선택): 연령대 필터 (예: \"20대\")\n- `keywordFilter` (기본 false): true이면 내 이상형 키워드와 일치하는 일기만\n\n**필터링 규칙:**\n- 이성 일기만 노출\n- 차단한/된 유저 제외\n- 이미 매칭 신청/넘긴 일기 제외\n- 교환일기 3건 만석 유저 제외\n\n**응답 카드:** authorId, ageGroupLabel, sido, sigungu, previewContent(앞 200자), personalityKeywords(상위3), moodTags",
        "operationId": "explore",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "latest"
            }
          },
          {
            "name": "sido",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sigungu",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ageGroup",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keywordFilter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": {
                    "diaries": [
                      {
                        "diaryId": 10,
                        "authorId": 5,
                        "ageGroupLabel": "20대",
                        "sido": "서울특별시",
                        "sigungu": "강남구",
                        "previewContent": "오늘은...",
                        "category": "DAILY",
                        "createdAt": "2026-04-30T10:00:00",
                        "similarityBadge": null,
                        "personalityKeywords": [
                          "안정 추구"
                        ],
                        "moodTags": [
                          "감성적인"
                        ]
                      }
                    ],
                    "nextCursor": 9,
                    "hasNext": true,
                    "guidanceMessage": null,
                    "currentSort": "latest"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}